+++ packages/abap-deploy-config-inquirer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/abap-deploy-config-inquirer
 
+## 1.6.30
+
+### Patch Changes
+
+-   Updated dependencies [4053369]
+    -   @sap-ux/inquirer-common@0.8.0
+
 ## 1.6.29
 
 ### Patch Changes