+++ packages/adp-flp-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sap-ux/adp-flp-config-sub-generator
 
+## 0.1.91
+
+### Patch Changes
+
+-   Updated dependencies [4053369]
+    -   @sap-ux/inquirer-common@0.8.0
+    -   @sap-ux/adp-tooling@0.16.4
+    -   @sap-ux/flp-config-inquirer@0.4.39
+
 ## 0.1.90
 
 ### Patch Changes