+++ packages/flp-config-inquirer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/flp-config-inquirer
 
+## 0.4.39
+
+### Patch Changes
+
+-   Updated dependencies [4053369]
+    -   @sap-ux/inquirer-common@0.8.0
+    -   @sap-ux/adp-tooling@0.16.4
+
 ## 0.4.38
 
 ### Patch Changes